TASK-4017 BUG: Whitespace after commas in chat message is lost
The protocol has no notion of a quoted string and instead sends the chat.post arguments as-is. The server splits on commas into individual arguments, and chat.post undoes that by re-joining with commas. With this approach, any whitespace after a comma gets lost.
Also pass allArgs to handleCommand() and directly use that for Server#broadcastMessage().