Build: Minor: Prevent warning about relative parent POM mismatch
When another pom.xml is in JuicyRaspberryPie's parent directory, a Maven build produces this warning:

> [WARNING] Some problems were encountered while building the effective model for org.wensheng:juicyraspberrypie:jar:1.18.1
> [WARNING] 'parent.relativePath' of POM org.wensheng:juicyraspberrypie:1.18.1 (/home/inkarkat/Resilient/JuicyRaspberryPie/pom.xml) points at works.reload:rebuild instead of works.reload:parent, please verify your project structure @ line 5, column 10
> [WARNING]
> [WARNING] It is highly recommended to fix these problems because they threaten the stability of your build.
> [WARNING]
> [WARNING] For this reason, future Maven versions might no longer support building such malformed projects.

Our MavenSetup repo (which provides the parent artifact) is typically checked out along JuicyRaspberryPie, but will never contain JuicyRaspberryPie. Therefore, the default relativePath is wrong, and should be cleared, so that the artifact is always resolved through the Maven repository.
Source: https://stackoverflow.com/a/6006098/813602