Build: Refactoring: Reuse MavenSetup generic Java build workflow
In order to tag and publish the built artifacts (JARs) to our Maven repository. The MavenSetup workflow does that already, and can be reused as-is.
Cp. https://docs.github.com/en/actions/using-workflows/reusing-workflows
This replaces the upload as GitHub artifacts; unlike the original repo, we haven't done any GitHub releases.
In fact, we didn't even touch the original version number so far; let's deviate from the original "following the compatible Minecraft release" versioning (which doesn't make sense for us and likely isn't a good idea in general) and instead restart with a 2.0.0 SemVer.
Our Maven repo is (for us) a more consistent and comfortable way of publishing (as we can use Maven version ranges and reference snapshot releases) than using GitHub releases.