TASK-4337 Generalization: Accept any kind of context object in the SessionAttachment
This isn't necessarily about EventQueues; any kind of data can be attached to a session. This also simplifies the method signatures at the cost of some casts in clients.
Registry#{start,stop}EventQueues() becomes Registry#createContexts().
The registration of the event listener is now done by the EventQueue constructor; for that, the plugin (and SessionAttachment because it might be useful in some cases, too) is passed in.
If the context object is auto-closeable, it will be closed when the Session is closed. That's where the unregistration of the event listener is now happening.
SessionAttachment itself is auto-closable, too; so the RemoteSession now only invokes attachment.close() and the SessionAttachment forwards the closing to any context objects that are auto-closeable.